refactor: extract shared cross-host subagent-followup formatter
Move the subagent follow-up XML block parsing, formatting, entity
decoding, and progress-detail logic into a host-neutral module at
src/shared/subagentFollowup.ts (no vscode, no Ink/React imports) so the
CLI transcript and the extension ProgressView bubble share one source of
truth. The CLI module re-exports the shared logic to keep its terse-line
output byte-for-byte identical and its ./subagentFollowup import path
stable; UserMessage.ts now consumes the shared decodeXmlEntities instead
of a duplicate decode path, keeping its own bubble/markdown styling and
its broader structured-delivery tag classification.
